src/main/webapp/app/entities/paper/detail/paper-detail.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { ActivatedRoute } from '@angular/router';
import { of } from 'rxjs';

import { DataUtils } from 'app/core/util/data-util.service';

import { PaperDetailComponent } from './paper-detail.component';

describe('Component Tests', () => {
  describe('Paper Management Detail Component', () => {
    let comp: PaperDetailComponent;
    let fixture: ComponentFixture<PaperDetailComponent>;
    let dataUtils: DataUtils;

    beforeEach(() => {
      TestBed.configureTestingModule({
        declarations: [PaperDetailComponent],
        providers: [
          {
            provide: ActivatedRoute,
            useValue: { data: of({ paper: { id: 123 } }) },
          },
        ],
      })
        .overrideTemplate(PaperDetailComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(PaperDetailComponent);
      comp = fixture.componentInstance;
      dataUtils = TestBed.inject(DataUtils);
      jest.spyOn(window, 'open').mockImplementation(() => null);
    });

    describe('OnInit', () => {
      it('Should load paper on init', () => {
        // WHEN
        comp.ngOnInit();

        // THEN
        expect(comp.paper).toEqual(expect.objectContaining({ id: 123 }));
      });
    });

    describe('byteSize', () => {
      it('Should call byteSize from DataUtils', () => {
        // GIVEN
        jest.spyOn(dataUtils, 'byteSize');
        const fakeBase64 = 'fake base64';

        // WHEN
        comp.byteSize(fakeBase64);

        // THEN
        expect(dataUtils.byteSize).toBeCalledWith(fakeBase64);
      });
    });

    describe('openFile', () => {
      it('Should call openFile from DataUtils', () => {
        // GIVEN
        jest.spyOn(dataUtils, 'openFile');
        const fakeContentType = 'fake content type';
        const fakeBase64 = 'fake base64';

        // WHEN
        comp.openFile(fakeBase64, fakeContentType);

        // THEN
        expect(dataUtils.openFile).toBeCalledWith(fakeBase64, fakeContentType);
      });
    });
  });
});
